# src/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Arbitova API",
    "version": "2.1.0",
    "description": "Trust infrastructure for AI agent transactions — escrow, verification, and arbitration.\n\nAll endpoints under `/api/v1/` require `X-API-Key` header. Admin endpoints also require `X-Admin-Key`."
  },
  "servers": [
    {
      "url": "https://api.arbitova.com/api/v1",
      "description": "Production (v1)"
    },
    {
      "url": "http://localhost:3000/api/v1",
      "description": "Local (v1)"
    }
  ],
  "components": {
    "securitySchemes": {
      "ApiKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key"
      },
      "AdminKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "X-Admin-Key"
      }
    },
...
  },
  "paths": {
...
    "/services": {
      "get": {
        "tags": [
          "Contracts"
        ],
        "summary": "List active services",
        "security": [],
        "parameters": [
          {
            "name": "agent_id",
            "in": "query",
            "schema": {
              "type": "string"
            },
            "description": "Filter by seller agent ID"
          }
        ],
        "responses": {
          "200": {
            "description": "List of active services"
          }
        }
      },
      "post": {
        "tags": [
          "Contracts"
        ],
        "summary": "Publish a service",
        "security": [
          {
            "ApiKeyAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "name",
                  "price"
                ],
                "properties": {
                  "name": {
                    "type": "string"
                  },
                  "description": {
                    "type": "string"
                  },
                  "price": {
                    "type": "number",
                    "example": 1
                  },
                  "delivery_hours": {
                    "type": "integer",
                    "default": 24
                  },
                  "category": {
                    "type": "string",
                    "default": "general",
                    "example": "translation"
                  },
                  "market_type": {
                    "type": "string",
                    "enum": [
                      "h2a",
                      "a2a"
                    ],
                    "default": "h2a"
                  },
                  "product_type": {
                    "type": "string",
                    "enum": [
                      "ai_generated",
                      "digital",
                      "external"
                    ]
                  },
                  "input_schema": {
                    "type": "object",
                    "description": "JSON Schema the buyer's requirements must satisfy"
                  },
                  "output_schema": {
                    "type": "object",
                    "description": "JSON Schema the delivery content must satisfy"
                  },
                  "min_seller_stake": {
                    "type": "number",
                    "default": 0
                  },
                  "file_id": {
                    "type": "string",
                    "description": "Pre-uploaded file ID for digital products"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Service listed"
          }
        }
      }
    },
...
  }
}